further migrate functions and rename modules
